# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "saitokdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/directions/address/{lat}/{lng}": {
            "get": {
                "tags": [
                    "Direction"
                ],
                "summary": "Get address from location",
                "operationId": "f397580c17dbf9319a07632926c9bd3d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/lat"
                    },
                    {
                        "$ref": "#/components/parameters/lng"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/directions/location/{address}": {
            "get": {
                "tags": [
                    "Direction"
                ],
                "summary": "Get location from address",
                "operationId": "373f1d2310d4483a3aa0035f41356144",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/address"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/notification/get/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get a notification",
                "operationId": "36e5cc47189824900663658258aac4fc",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/notification_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/notification/app/{page}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get all notification",
                "operationId": "6fc9c48ccdddecc0b6367746efaf1e1f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/notification/read/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Read notification",
                "operationId": "0b04247dbc18091d7fd419af03bdfe8f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/notification_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/notification/types/{type}/{page}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get Separated Notifications",
                "operationId": "6d9f6441cca8fbd7c5e0742a0434809e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/type"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/notification/delete/{id}": {
            "delete": {
                "tags": [
                    "Notification"
                ],
                "summary": "Delete a notification",
                "operationId": "3dee53d70e75ac52cea5db72e6a16869",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/notification_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/notification/unread": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get unread count",
                "operationId": "ee74ee5169aee760e08037a9992f0cd7",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
...
            "lat": {
                "name": "lat",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
            "lng": {
                "name": "lng",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
            "address": {
                "name": "address",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "type": {
                "name": "type",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "page": {
                "name": "page",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "notification_id": {
                "name": "notification_id",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}